dom/tests/mochitest/bugs/test_bug597809.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=597809
-->
<head>
  <title>Test for Bug 597809</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=597809">Mozilla Bug 597809</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 597809 **/

SimpleTest.waitForExplicitFinish();

SpecialPowers.addCategoryEntry("JavaScript-global-property", "testSNSM", "@mozilla.org/xmlextras/xmlhttprequest;1",
                    false, true);

SimpleTest.executeSoon(function () {
  ok(window.testSNSM, "testSNSM should return an object");
  // The category entry must be removed before finishing the test,
  // otherwise it will affect all following tests in the test suite.
  SpecialPowers.deleteCategoryEntry("JavaScript-global-property", "testSNSM", false);
  SimpleTest.finish();
});


</script>
</pre>
</body>
</html>
